Development/Backend Development/Database Connections/requirements.txt
boto3>=1.26.0
mysql-connector-python>=8.0.0
psycopg2-binary>=2.9.0
pymongo>=4.0.0
redis>=4.0.0
cx_Oracle>=8.0.0
sqlalchemy>=2.0.0
python-dotenv>=0.19.0
cassandra-driver>=3.25.0
neo4j>=4.4.0
